requirements.txt
streamlit==1.31.0
Pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
unstructured==0.12.0
unstructured-inference==0.7.23
pytesseract==0.3.10
pdf2image==1.17.0
python-magic==0.4.27